Accordionのコンポーネント化にあたりリネーム。
