+++ src/components/molecules/MonthlyArchive.stories.tsx
@@ -1,17 +1,17 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import ArticlesCount from "../../api/ArticlesCountMockDate.json";
 
-import { Accordion } from "./Accordion";
+import { MonthlyArchive } from "./MonthlyArchive";
 
 const meta = {
     title: "Accordion",
-    component: Accordion,
+    component: MonthlyArchive,
     parameters: {
         layout: "centered",
     },
     tags: [ "autodocs" ],
     argTypes: {},
-} satisfies Meta<typeof Accordion>;
+} satisfies Meta<typeof MonthlyArchive>;
 
 export default meta;
